+++ autoload.php
@@ -31,8 +31,8 @@ require_once __DIR__.'/core/LibSFTP.php';
 require_once __DIR__ . '/core/LibSession.php';
 
 require_once __DIR__.'/SmallPHPMail/phpmailerException.php';
-require_once __DIR__.'/SmallPHPMail/class.phpmailer.php';
-require_once __DIR__.'/SmallPHPMail/class.smtp.php';
+require_once __DIR__.'/SmallPHPMail/PHPMailer.php';
+require_once __DIR__.'/SmallPHPMail/SMTP.php';
 
 require_once __DIR__.'/core/LibMail.php';